s3/requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
fallocate==1.6.4
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
leval==1.2.0
psycopg2==2.9.9
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
urllib3==2.2.0
valohai-papi==0.1.3
valohai-utils==0.4.0
valohai-yaml==0.38.0
